[CURL] Access CURL object via `CFileItem::GetURL` and `CFileItem::GetDynURL`.
These methods will be updated in a later branch to be very cheap to call.

As it is now possible to cheaply acquire a `CURL` object from a
`CFileItem` object it will be more performant to add utility methods
onto the CURL object. These methods will be better adapted than the
`URIUtils::*` functions.